Use a descendant combinator for a descendant
The descendant combinator is the right choice when applying styles to
elements that are descendants of a specific element. Thereby, it leaves
elements that are not descendants of the respective element unstyled.

Type or class selectors can be used for the descendant combinator or
combined together, just like I did it here by combining the type
selector "div" and the class selector "text".