Modify padding and margin of all three objects
Note that I used "margin-left: auto" in order to align the third object
on the right side with the second object. Generally, the value
"margin-left: auto" receives a share of the unused horizontal space.

Using both margin-left and margin-right and setting the values of both
to auto would distribute the calculated space evenly, eventually leading
to the centering of an object.